Move silence functions
Move silence-related functions into their own package for modularity.
